backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
sqlmodel
pydantic
pydantic-settings
python-dotenv
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
celery
redis
feedparser
beautifulsoup4
newspaper3k
openai
google-generativeai
cryptography
httpx
requests
python-docx
